Threads/requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
joblib==1.5.1
msgpack==1.1.1
numpy==2.3.1
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
redis==6.2.0
scikit-learn==1.7.0
scipy==1.16.0
shortuuid==1.0.13
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
